refacto Welcome component
