+++ src/components/Dashboard/Welcome/welcome.scss
@@ -1,4 +1,4 @@
-@import 'src/styles/vars';
+@import '@/styles/vars';
 
 #welcomeBoardMessage{
   animation-duration: 0.3s;
