--- src/containers/Dashboard/Welcome.js
@@ -1,25 +1,0 @@
-// == Import : npm
-import { connect } from 'react-redux';
-
-// == Import : local
-import Welcome from 'src/components/Dashboard/Welcome';
-
-// Action Creators
-
-
-/* === State (données) === */
-const mapStateToProps = (state) => ({
-//   message: state.appReducer.message,
-});
-
-/* === Actions === */
-const mapDispatchToProps = {};
-
-// Container
-const WelcomeContainer = connect(
-  mapStateToProps,
-  mapDispatchToProps,
-)(Welcome);
-
-// == Export
-export default WelcomeContainer;
